+++ packages/react-magma-dom/src/index.ts
@@ -217,5 +217,6 @@ export { I18nInterface } from './i18n/interface';
 export { useMediaQuery } from './hooks/useMediaQuery';
 export { InverseContext, useIsInverse } from './inverse';
 
+export * from './components/Accordion';
 export * from './components/Transition';
 export * from './components/Drawer';